+++ public/assets/picket-sign/picket-sign-flipped.png
