+++ public/assets/picket-sign/picket-sign.png
