+++ public/assets/sign-content/contract-spelling-bee.png
