+++ public/assets/sign-content/just-cause-wordle.png
