+++ src/types.ts
@@ -1,8 +1,8 @@
 export enum SignContent {
-  ContractSpellingBee,
-  GuildRound,
-  GuildSquare,
-  JustCauseWordle,
-  MakerWeekGuild,
-  UnionMadeGithub
+  ContractSpellingBee = 'contract-spelling-bee.png',
+  GuildRound = 'guild-logo-round.png',
+  GuildSquare = 'guild-logo-square.png',
+  JustCauseWordle = 'just-cause-wordle.png',
+  MakerWeekGuild = 'maker-week-guild.png',
+  UnionMadeGithub = 'union-made-github.png'
 }
